requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
imagekitio
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
